refactor: migrate config imports to use unified central @/config module
Update import paths across DashboardLayout, next-auth, and middleware files.
Consolidate configuration imports from scattered individual lib files to the central @/config barrel export. This standardizes import patterns, reduces duplication, and simplifies future config module maintenance.
