+++ src/config/index.ts
@@ -1,0 +1,11 @@
+/**
+ * Application Configuration
+ * Centralized access to all system configuration settings
+ */
+
+export * from "./session.config";
+export * from "./security.config";
+
+// Re-export for default imports
+export { default as SESSION_CONFIG } from "./session.config";
+export { securityConfig } from "./security.config";
