+++ src/middleware.ts
@@ -1,7 +1,7 @@
 import { NextResponse } from "next/server";
 import type { NextRequest } from "next/server";
 import { auth } from "@/lib/next-auth";
-import { validateSessionSecurity } from "@/lib/security";
+import { validateSessionSecurity } from "@/config";
 
 export async function middleware(request: NextRequest) {
   const { pathname } = request.nextUrl;
